+++ src/__test__/redux/actions/projects/updateMetadataTrack.test.js
@@ -1,6 +1,6 @@
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { metadataNameToKey } from '../../../../utils/metadataUtils';
+import { metadataNameToKey } from '../../../../utils/data-management/metadataUtils';
 import updateMetadataTrack from '../../../../redux/actions/projects/updateMetadataTrack';
 import initialProjectState from '../../../../redux/reducers/projects';
 import initialSampleState from '../../../../redux/reducers/samples';
