+++ src/__test__/utils/metadataUtils.test.js
@@ -2,7 +2,7 @@ import {
   metadataKeyToName,
   metadataNameToKey,
   temporaryMetadataKey,
-} from '../../utils/metadataUtils';
+} from '../../utils/data-management/metadataUtils';
 
 describe('metadataUtils', () => {
   it('metadataKeyToName converts name correctly', () => {
