+++ src/components/data-management/MetadataPopover.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Popover } from 'antd';
 import EditableField from '../EditableField';
-import { metadataKeyToName, metadataNameToKey } from '../../utils/metadataUtils';
+import { metadataKeyToName, metadataNameToKey } from '../../utils/data-management/metadataUtils';
 import validateInputs, { rules } from '../../utils/validateInputs';
 
 const validationChecks = [
