+++ src/redux/actions/projects/updateMetadataTrack.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { metadataNameToKey } from '../../../utils/metadataUtils';
+import { metadataNameToKey } from '../../../utils/data-management/metadataUtils';
 import {
   PROJECTS_METADATA_UPDATE,
 } from '../../actionTypes/projects';
